third_party/blink/web_tests/fast/xsl/generate-id-multiple-calls-same-value-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
Verify that generate-id() returns the same value each time it is called for a given node in the current document in the current transform.

Value 1 IDs matched as expected: PASSED

Value 2 IDs matched as expected: PASSED

Value 3 IDs matched as expected: PASSED
